+++ e2e/cypress/dataHooks.js
@@ -14,6 +14,7 @@ export const INLINE_TOOLBAR_BUTTONS = {
   TEXT_ALIGN_CENTER: 'textAlignmentButton_center',
   TEXT_ALIGN_LEFT: 'textAlignmentButton_left',
   TEXT_ALIGN_RIGHT: 'textAlignmentButton_right',
+  TEXT_ALIGN_JUSTIFY: 'textAlignmentButton_justify',
   LINE_SPACING: 'LineSpacingButton',
   LINK: 'LinkButton',
   CODE_BLOCK: 'TextCodeBlockButton',