+++ e2e/tests/text.e2e.js
@@ -1,7 +1,7 @@
 /* eslint-disable max-len */
 /*global cy*/
 import { INLINE_TOOLBAR_BUTTONS } from '../cypress/dataHooks';
-import { DEFAULT_DESKTOP_BROWSERS } from './settings';
+import { DEFAULT_DESKTOP_BROWSERS, FIREFOX_BROWSER } from './settings';
 import { usePlugins, plugins } from '../cypress/testAppConfig';
 
 describe('text', () => {
@@ -252,5 +252,46 @@ describe('text', () => {
         .blurEditor();
       cy.eyesCheckWindow(this.test.title);
     });
+  });
+});
+
+describe('textFirefox', { env: { firefox: true } }, () => {
+  before(function() {
+    cy.eyesOpen({
+      appName: 'textFirefox',
+      testName: this.test.parent.title,
+      browser: FIREFOX_BROWSER,
+    });
+  });
+
+  beforeEach(() => cy.switchToDesktop());
+
+  afterEach(() => cy.matchContentSnapshot());
+
+  after(() => cy.eyesClose());
+
+  it('Enter click and space should keep the line when justified', function() {
+    cy.loadRicosEditorAndViewer()
+      .enterParagraphs([
+        '   Leverage agile frameworks to provide a robust synopsis for high level overviews. Iterative approaches to corporate strategy foster. ',
+      ])
+      .setEditorSelection(0, 134)
+      .setAlignment(INLINE_TOOLBAR_BUTTONS.TEXT_ALIGN_JUSTIFY)
+      .setEditorSelection(134, 0)
+      .type('{enter}')
+      .type('{enter}')
+      .enterParagraphs([' '])
+      .type('{enter}')
+      .enterParagraphs(['next line'])
+      .blurEditor();
+    cy.eyesCheckWindow(this.test.title);
+  });
+
+  it('allow to create justify lists', function() {
+    cy.loadRicosEditorAndViewer('plain', {}, true)
+      .setTextStyle(INLINE_TOOLBAR_BUTTONS.ORDERED_LIST, [300, 100])
+      .setAlignment(INLINE_TOOLBAR_BUTTONS.TEXT_ALIGN_JUSTIFY)
+      .blurEditor();
+    cy.eyesCheckWindow(this.test.title);
   });
 });